requirements.txt
hydra-core
torch
torchvision
numpy
pandas
matplotlib
opencv-python
#scipy
tqdm
#scikit-image
#tensorboard
#sklearn
argparse
pytorch-msssim